requirements.txt
python-dotenv
Flask
OpenAI
llama_index
langchain